refactor | HTMLElements: move into containers dir
