+++ src/containers/MaginPreview/MaginPreview.tsx
